+++ routes/commentaires.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const router = express.Router();
+const auth = require("../middleware/auth");
 const {
   getAllCommentaires,
   getCommentaireById,
@@ -7,6 +8,8 @@ const {
   updateCommentaire,
   deleteCommentaire,
 } = require("../models/commentaires");
+const { PrismaClient } = require("@prisma/client");
+const prisma = new PrismaClient();
 
 //get all commentaires
 router.get("/", (req, res, next) => {
@@ -21,19 +24,34 @@ router.get("/:id([0-9]+)", (req, res, next) => {
 });
 
 //post commentaire
-router.post("/", (req, res, next) => {
+router.post("/", auth, (req, res, next) => {
+  if (req.user.iduser != req.body.iduser)
+    return res.status(401).send("operation not allowed");
+
   addCommentaire(req.body).then((commentaire) => res.json(commentaire));
 });
 
 //put commentaire
-router.put("/:id([0-9]+)", (req, res, next) => {
+router.put("/:id([0-9]+)", auth, (req, res, next) => {
+  if (req.user.iduser != req.body.iduser)
+    return res.status(401).send("operation not allowed");
+
   updateCommentaire(req.params.id, req.body).then((commentaire) =>
     res.json(commentaire)
   );
 });
 
 //delete commentaire
-router.delete("/:id([0-9]+)", (req, res, next) => {
+router.delete("/:id([0-9]+)", auth, async (req, res, next) => {
+  const comment = await prisma.commentaire.findUnique({
+    where: {
+      idcommentaire: +req.params.id,
+    },
+  });
+  if (!comment) return res.status(404).send("comment not found");
+  if (req.user.iduser != comment.iduser)
+    return res.status(401).send("operation not allowed");
+
   deleteCommentaire(req.params.id).then((commentaire) => res.json(commentaire));
 });
